New API endpoints for cloudinary image hosting. Save Album function needs to be adjusted

+++ src/components/ImageList.js
@@ -6,41 +6,38 @@ import { connect } from 'react-redux'
 import { changeOrder } from '../actions/photos-actions'
 
 export class ImageList extends Component {
-    state = {
-        newList: []
-    }
 
     constructor(props) {
         super(props)
-        this.onchangeOrder = this.onchangeOrder.bind(this);
+        this.onSort = this.onSort.bind(this);
     }
-    onchangeOrder(newList) {
-        this.props.onchangeOrder(newList);
+    
+    onSort(sortedList) {
+        let newArray = []
+        sortedList.forEach(image => {
+            newArray.push(image.content.props.item)
+        })
+        this.props.onchangeOrder(newArray); 
     }
+
+
     render() {
-        // <div className="image-lists">
-        //     <div className="image-lists__portraits">
-        // {(this.props.photos.map((item, index) =>
-        //     <ImageBlade item={item} key={index} index={index} category={this.props.category} deleteItem={this.props.deleteItem}/>
-        // ))}
-        //     </div>
-        // </div>
-        var onSort = function(sortedList) {
-            console.log("sortedList", sortedList);
-        }
         var list = []
-        this.props.photos.forEach((item, index) => {
-            list.push({
-                content: (<ImageBlade item={item} key={index} index={index} category={this.props.category} deleteItem={this.props.deleteItem}/>),
-                classes: ['image-list__photo']
-            })
-        });
+        if(this.props.photos.length > 0) {
+            this.props.photos.map((item, index) => {
+                list.push({
+                    content: (<ImageBlade item={item} key={index} index={index} category={this.props.category} deleteItem={this.props.deleteItem}/>),
+                    classes: ['image-list__photo']
+                })
+            });
+        }
 
         return (
             <div className="image-list">
-                <DragSortableList items={list} moveTransitionDuration={0.3} onSort={onSort} type="vertical"/>
+                <DragSortableList items={list} moveTransitionDuration={0.3} onSort={this.onSort} type="vertical"/>
             </div>
         ) 
+
     }
 }
